Removed idOwners from pet tables
Now the name and surname of the person who owns the pet is displayed instead of its id.
To do this, we have included the getPeople method in the people dao to obtain the information of a specific person and thus obtain the name and surname to be shown in the table.