Changes in the form of pets
The list of existing people in a pet owner field has been added to better identify the person we want to select.
In addition, it has been added that in the view of a person's pets, the name of that person appears by default in the form of adding / editing the information of a pet.